bitcoin-sentiment-tracker-demo/backend/requirements.txt
Flask>=2.3
feedparser>=6.0
google-generativeai>=0.5